depId, posId int에서 Long으로 수정완료

+++ src/main/java/com/kh/coreflow/model/dto/UserDto.java
@@ -41,8 +41,8 @@ public class UserDto {
 		private String email;
 		private String userPwd;
 		private List<String> roles;
-		private int depId;
-		private int posId;
+		private Long depId;
+		private Long posId;
 		private Date hireDate;
 		private String extension;
 		private String phone;
